+++ src/withFacets.js
@@ -24,6 +24,7 @@ import {
   browseModeMap,
   browseCallNumberOptions,
   queryIndexes,
+  browseClassificationOptions,
 } from './constants';
 import { getAdvancedSearchTemplate } from './routes/buildManifestObject';
 
@@ -125,6 +126,10 @@ function withFacets(WrappedComponent) {
 
       if (queryIndex === browseModeOptions.SUBJECTS) {
         return 'search/subjects/facets';
+      }
+
+      if (Object.values(browseClassificationOptions).includes(queryIndex)) {
+        return 'search/classifications/facets';
       }
 
       return 'search/instances/facets';